Upgrading cargo packages to newest versions. Fixes for broken code due to upgrades.

+++ src/database/sqlite.rs
@@ -546,7 +546,7 @@ impl super::Database for SQLite {
                     Ok(val) => {
                         if val > 0 {
                             let mut tmp = r.clone();
-                            tmp.set_id(tx.last_insert_rowid() as u64);
+                            tmp.set_id(tx.last_insert_rowid());
                             output.insert(0, tmp);
                         }
                     },
